+++ Model/Attribute/DefaultValueProvider/AbstractProvider.php
@@ -46,6 +46,13 @@ abstract class AbstractProvider extends \Magento\Framework\DataObject
     public function canBeUsedForAttribute()
     {
         $applicatbleInputTypes =  $this->getApplicableFrontendInputs() ?? [];
+
+        if ($this->attribute->getSourceModel()) {
+            if (!$this->attribute->getSourceModel()->canUseDefaultValue()) {
+                return false;
+            }
+        }
+
         if (!in_array($this->attribute->getFrontendInput(), $applicatbleInputTypes)) {
            return false;
         }